+++ R/verb-filter.R
@@ -53,8 +53,8 @@ add_filter <- function(.data, dots) {
     return(filter_via_having(lazy_query, dots))
   }
 
-  if (!uses_window_fun(dots, con)) {
-    if (uses_mutated_vars(dots, lazy_query$select)) {
+  if (!dots_use_window_fun) {
+    if (filter_needs_new_query(dots, lazy_query, con)) {
       lazy_select_query(
         x = lazy_query,
         where = dots
@@ -86,6 +86,26 @@ add_filter <- function(.data, dots) {
       where = where$expr
     )
   }
+}
+
+filter_needs_new_query <- function(dots, lazy_query, con) {
+  if (!inherits(lazy_query, "lazy_select_query")) {
+    return(TRUE)
+  }
+
+  if (uses_mutated_vars(dots, lazy_query$select)) {
+    return(TRUE)
+  }
+
+  if (uses_window_fun(lazy_query$select$expr, con)) {
+    return(TRUE)
+  }
+
+  if (any_expr_uses_sql(lazy_query$select$expr)) {
+    return(TRUE)
+  }
+
+  FALSE
 }
 
 filter_can_use_having <- function(lazy_query, dots_use_window_fun) {